Added minuteConversion;

+++ variablesAndDataTypes/operators/minuteConversion/minuteConversion.c
@@ -1,16 +1,16 @@
 #include <stdio.h>
 
 int main(){
-    const int minutesInYear = 60 * 24 * 365; // minutesinYour * hoursInDay * daysInYear ~= minutesInYear
-    const int minutesInDay = 60 * 24; // minutesinYour * hoursInDay ~= minutesInDay
+    const double minutesInYear = 60 * 24 * 365; // minutesinYour * hoursInDay * daysInYear ~= minutesInYear
+    const double minutesInDay = 60 * 24; // minutesinYour * hoursInDay ~= minutesInDay
     int minutes;
     double years, days;
     
     printf("Insert the number of minutes to convert: ");
     scanf("%d", &minutes);
     
-    days = (double)minutes / (double)minutesInDay;
-    years = (double)minutes / (double)minutesInYear;
+    days = minutes / minutesInDay;
+    years = minutes / minutesInYear;
     
     printf("Days: %lf\n", days);
     printf("Years: %lf\n", years);
